added update & delete operations for Country entity ./

+++ src/main/java/com/taahaagul/ifiwastemanagement/service/CountryService.java
@@ -1,9 +1,11 @@
 package com.taahaagul.ifiwastemanagement.service;
 
 import com.taahaagul.ifiwastemanagement.entity.Country;
+import com.taahaagul.ifiwastemanagement.exception.IllegalOperationException;
 import com.taahaagul.ifiwastemanagement.exception.ResourceNotFoundException;
 import com.taahaagul.ifiwastemanagement.repository.CountryRepository;
 import com.taahaagul.ifiwastemanagement.request.CountryRequest;
+import com.taahaagul.ifiwastemanagement.request.CountryUpdateRequest;
 import com.taahaagul.ifiwastemanagement.response.CountryResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -40,6 +42,20 @@ public class CountryService {
         Country foundedCountry = countryRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Country", "id", id.toString()));
 
+        if (!foundedCountry.getCities().isEmpty()) {
+            throw new IllegalOperationException("Cannot delete Country as it is associated with one or more City entities");
+        }
+
         countryRepository.delete(foundedCountry);
+    }
+
+    public CountryResponse updateCountry(Long id, CountryUpdateRequest countryUpdateRequest) {
+        Country foundedCountry = countryRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Country", "id", id.toString()));
+
+        foundedCountry.setCountryName(countryUpdateRequest.getCountryName());
+        foundedCountry.setCounrtyCode(countryUpdateRequest.getCountryCode());
+
+        return new CountryResponse(countryRepository.save(foundedCountry));
     }
 }
